ms/category_ms/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
circuitbreaker==2.0.0
click==8.1.7
Flask==3.0.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
logzero==1.7.0
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.20.1
marshmallow-sqlalchemy==1.0.0
packaging==24.1
proofdock-chaos-middleware-flask==0.1.3
proofdock-chaos-middleware-python==0.2.3
psycopg2==2.9.7
python-dotenv==0.19.0
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.30
sqlalchemy-filters==0.13.0
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
